+++ backend/lobby/manager.py
@@ -22,10 +22,13 @@ class Manager:
 	# 	now = datetime.now()
 	# 	with self._lock:
 	# 		self._lobbies = {code: lobby for code, lobby in self._lobbies.items() if now < lobby.created + self._inative_timeout}
-	def _on_game_finish(self, history: GameHistory, lobby: Lobby):
+	def close_lobby(self, lobby: Lobby):
 		with self._lock:
 			del self._lobbies[lobby.code]
-			self.db.queue_commit_game_history(history)
+			self._player_code = {id: code for id, code in self._player_code.items() if code != lobby.code}
+
+	def _on_game_finish(self, history: GameHistory):
+		self.db.queue_commit_game_history(history)
 
 	def create_lobby(self, user: User, code_length: int = 5) -> str:
 		code = "".join(choices(ascii_uppercase, k=code_length))
@@ -34,11 +37,16 @@ class Manager:
 				raise Exception("lobby limit reached")
 			elif code in self._lobbies:
 				raise Exception("bad luck")
-			elif user.user_id in self._player_code:
+			elif self.__in_lobby(user):
 				raise Exception("player already in lobby")
-			self._lobbies[code] = Lobby(code, self._on_game_finish)
+			self._lobbies[code] = Lobby(code, user.user_id, self._on_game_finish)
 			self.join_lobby(user, code)
 		return code
+
+	def __in_lobby(self, user: User) -> bool:
+		try: self._lobbies[self._player_code[user.user_id]]
+		except: return False
+		return True
 
 	def join_lobby(self, user: User, code: str) -> None:
 		with self._lock: