+++ frontend/src/settings/index.ts
@@ -1,4 +1,3 @@
 export { default as SettingsProvider, useSettings } from "./SettingsProvider"
-export { default as SettingsDialog } from "./SettingsDialog"
 export * from "./types"
 export { default as useNotification } from "./useNotification"